requirements.txt
# Core Framework
streamlit==1.28.0
langgraph==0.0.20
langchain==0.1.0
langchain-community==0.0.10

# LLM Integration
ollama==0.1.6

# Web Scraping
selenium==4.15.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Task Scheduling
apscheduler==3.10.4

# Security
cryptography==41.0.7

# Notifications
twilio==8.10.0

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
